ADD: Data templates menu entry
The Germinate platform supports importing data via template files; however, these template files are not bundled with the platform itself. Instead, they are hosted externally at https://github.com/germinateplatform/germinate-data-templates.

Because the location of these files lies outside the platform’s base URL set according the server where it is installed and users need a consistent, familiar place to find and download them, this change introduces a dedicated menu entry. This menu item redirects users directly to the download location for the templates.

The redirect URL is configurable via the internationalization variable `menuDataTemplatesURL`. If this variable is missing or empty or not exists in the translation file, the menu defaults to redirecting users to the official GitHub repository page.